server/pyproject.toml
[tool.poetry]
name = "ednaldo"
version = "0.1.0"
description = ""
authors = ["Ednaldo"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-wtf = "^1.1.1"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.5"
psycopg2 = "^2.9.7"
load-dotenv = "^0.1.0"
flask-migrate = "^4.0.4"
boto3 = "^1.28.48"
flask-hashids = "^1.0.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
